backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Todo Full-Stack Web Application Backend"
authors = [
    {name = "Developer", email = "developer@example.com"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sqlmodel>=0.0.16",
    "pydantic[email]>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",
    "psycopg[binary,pool]>=3.1.15",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0"
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "isort>=5.10.0",
    "flake8>=6.0.0"
]